Refactor name of file
